+++ packages/bako-safe/tsup.config.js
@@ -4,5 +4,6 @@ import { defineConfig } from 'tsup';
 export default defineConfig((options) => ({
   ...baseConfig(options, { withReact: false }),
   format: ['cjs'],
+  external: ['fuels', 'socket.io-client'],
   entry: ['src/index.ts'],
 }));