+++ packages/connectors/src/defaultConnectors.ts
@@ -1,5 +1,6 @@
 import { BakoSafeConnector } from '@fuel-connectors/bako-safe';
 import { EVMWalletConnector } from '@fuel-connectors/evm-connector';
+import { BurnerWalletConnector } from '@fuel-connectors/burner-wallet-connector';
 import { FuelWalletDevelopmentConnector } from '@fuel-connectors/fuel-development-wallet';
 import { FuelWalletConnector } from '@fuel-connectors/fuel-wallet';
 import { FueletWalletConnector } from '@fuel-connectors/fuelet-wallet';
@@ -17,6 +18,7 @@ export function defaultConnectors({
     new FueletWalletConnector(),
     new BakoSafeConnector(),
     new EVMWalletConnector(),
+    new BurnerWalletConnector(),
   ];
 
   if (devMode) {