+++ packages/connectors/tsup.config.js
@@ -3,14 +3,20 @@ import { defineConfig } from 'tsup';
 
 export default defineConfig((options) => ({
   ...baseConfig(options, { withReact: false }),
-  platform: 'browser',
-  entry: ['src/index.ts'],
+  entry: [
+    'src/index.ts',
+    'src/walletconnect-connector/index.ts',
+    'src/evm-connector/index.ts',
+  ],
   external: ['fuels'],
   noExternal: [
     '@fuel-connectors/fuel-development-wallet',
     '@fuel-connectors/fuel-wallet',
     '@fuel-connectors/fuelet-wallet',
-    '@fuel-connectors/bako-safe',
+    '@fuel-connectors/burner-wallet-connector',
+    '@fuel-connectors/evm-connector',
+    '@fuel-connectors/walletconnect-connector',
+    '@fuel-connectors/bakosafe-connector',
   ],
   minify: 'terser',
   dts: {
@@ -18,7 +24,10 @@ export default defineConfig((options) => ({
       '@fuel-connectors/fuel-development-wallet',
       '@fuel-connectors/fuel-wallet',
       '@fuel-connectors/fuelet-wallet',
-      '@fuel-connectors/bako-safe',
+      '@fuel-connectors/burner-wallet-connector',
+      '@fuel-connectors/evm-connector',
+      '@fuel-connectors/walletconnect-connector',
+      '@fuel-connectors/bakosafe-connector',
     ],
   },
   splitting: true,