+++ packages/evm-connector/src/constants.ts
@@ -7,4 +7,4 @@ declare global {
 }
 const HAS_WINDOW = typeof window !== 'undefined';
 export const WINDOW = HAS_WINDOW ? window : null;
-export const BETA_5_URL = 'https://beta-5.fuel.network/graphql';
+export const TESTNET_URL = 'https://testnet.fuel.network/v1/graphql';