+++ packages/walletconnect-connector/src/index.ts
@@ -1,0 +1,2 @@
+export * from './WalletConnectConnector';
+export * from './types';